writeSilence

+++ src/AudioTools/AudioOutput.h
@@ -449,13 +449,13 @@ public:
     stream_idx = idx;
   }
 
-  size_t writeSilence(int bytes) override {
+  size_t writeSilence(size_t bytes) override {
     if (bytes == 0) return 0;
     byte silence[bytes] = {0};
     return write(stream_idx, silence, bytes);
   }
 
-  size_t writeSilence(int idx, int bytes){
+  size_t writeSilence(int idx, size_t bytes){
     if (bytes == 0) return 0;
     byte silence[bytes] = {0};
     return write(idx, silence, bytes);
